requirements/requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
#Flask-SQLAlchemy
peewee
pyyaml
psycopg2
pymysql
